app_scratch/requirements.txt
fastapi==0.111.1
mangum==0.17.0
protobuf==5.27.2
python-dotenv==1.0.1
qdrant_client==1.10.1
